Remove "spisemisu/elm-utf8" from "tests/elm-package.json" and "example/elm-package.json".
Bump version from 2.1.0 to 2.1.1.

Everything works now, but we're left with our own copy of UTF8.elm. I will talk on the Elm Slack channel about how to fix this correctly. We may get more of it with people angry about Microsoft's acquisition of GitHub.

I may choose to fix it by creating "billstclair/elm-utf8", and encouraging users of
"spisemisu/elm-utf8" to switch.